feat: Implement RelationIdentification
