+++ lib/universal_id.rb
@@ -17,5 +17,7 @@ require_relative "universal_id/version"
 require_relative "universal_id/errors"
 require_relative "universal_id/config"
 require_relative "universal_id/portable"
+require_relative "universal_id/hydration"
 require_relative "universal_id/portable_hash"
 require_relative "universal_id/active_model_serializer"
+require_relative "universal_id/relation_identification"
