+++ lib/universal_id/portable_hash.rb
@@ -2,6 +2,7 @@
 
 class UniversalID::PortableHash < Hash
   include UniversalID::Portable
+  extend UniversalID::Hydration
 
   class << self
     def config
@@ -13,71 +14,6 @@ class UniversalID::PortableHash < Hash
       hydrate JSON.parse(Zlib::Inflate.inflate(compressed_json))
     rescue => error
       raise UniversalID::LocatorError.new(id, error)
-    end
-
-    def dehydrate(hash, options)
-      include_list = options[:only] || []
-      exclude_list = options[:except] || []
-      hash.each_with_object({}) do |(key, value), memo|
-        key = key.to_s
-        next if include_list.any? && include_list.none?(key)
-        next if exclude_list.any?(key)
-        deep_dehydrate(value, options: options) { |val| memo[key] = val }
-      end
-    end
-
-    alias_method :deep_transform, :dehydrate
-    UniversalID.deprecator.deprecate_methods self, :deep_transform, deep_transform: "Use `dehydrate` instead."
-
-    def hydrate(hash)
-      hash.each_with_object({}) do |(key, value), memo|
-        deep_hydrate(value) { |val| memo[key] = val }
-      end
-    end
-
-    private
-
-    def deep_dehydrate(value, options:)
-      value = if implements_gid?(value)
-        value.to_gid_param
-      elsif defined?(ActiveRecord) && value.is_a?(ActiveRecord::Associations::CollectionProxy)
-        # fall back to target relation
-        deep_dehydrate(value.scope, options: options)
-      elsif defined?(ActiveRecord) && value.is_a?(ActiveRecord::Relation)
-        deep_dehydrate(value.select(value.primary_key).to_a, options: options)
-      else
-        case value
-        when Array then value.map { |val| deep_dehydrate(val, options: options) }
-        when Hash then dehydrate(value, options)
-        else value
-        end
-      end
-
-      if block_given?
-        yield value if value.present? || options[:allow_blank]
-      end
-
-      value
-    end
-
-    def deep_hydrate(value)
-      value = if possible_gid_string?(value)
-        parse_gid(value) || value
-      else
-        case value
-        when Array then value.map { |val| deep_hydrate(val) }
-        when Hash then hydrate(value)
-        else value
-        end
-      end
-
-      value = value.find if value.is_a?(GlobalID)
-      yield value if block_given?
-      value
-    end
-
-    def implements_gid?(value)
-      value.respond_to? :to_gid_param
     end
   end
